Second update
# Important changes
1. Use Android AlertDialog naming style

|Before|After|
|::|::|
|yesButton|positiveButton|
|noButton|negativeButton|
|content|message|

2. If positive button title has been set to a string that is empty, or
only have whitespace and/or `\n`, it will be changed to `OK` (which
seems to be better than “AGREE”) to make sure, user can see the button
for dismissing the LLDialog.
3. Changed sequence of parameter of `setXXXButton`
4. Use optional type `Selector?` instead of `String` for action to
match the new style of swift
5. Allow value of title, message and title of negative button to be
nil, set them to be nil by default. That means these parameters can be
ignored.

6. Use function `show()` to show LLDialog instead of two steps
7. Response to rotation of device
8. Shortened closure expression inside function `disappear()`

# Other changes
9. Changed comments according to Markup Formatting Reference to display
a better quick help
10. Update `README.md` to match the function names
11. Reference `LLDialog.swift` to be the file inside folder `Source` so
next time only have to update one file
12. Since currently there is no modification in `Tests` or `UITests`,
these two targets and their files were removed
13. Removed overriding function `viewDidLoad()` in
“ViewController.swift”
13. Add run script to update build number when the project builds (Can
be found in Targets -> LLDialog -> Build Phases -> Run Script)
